Simplify
Write the function called simplify, which makes the denominator and denominator values of the simple fraction sent to it as simple as possible, together with a main function where the denominator and denominator values are entered from the keyboard. Print the values of the variables that hold the numerator and denominator values entered from the keyboard before sending them to the function and the changed states after they are sent to the screen. (Global variable should not be used and should also return 1 if the simplify function was simplified and zero (0) if it could not be done.)